+++ lib/datadog_flutter.dart
@@ -1,3 +1,6 @@
+import 'dart:convert';
+import 'dart:io';
+
 import 'package:datadog_flutter/src/tracking_consent.dart';
 import 'package:datadog_flutter/src/channel.dart';
 export 'package:datadog_flutter/src/tracking_consent.dart';
@@ -49,7 +52,7 @@ class DatadogFlutter {
     return await channel.invokeMethod('setUserInfo', {
       'id': id,
       'email': email,
-      'extraInfo': extraInfo,
+      'extraInfo': Platform.isIOS ? jsonEncode(extraInfo) : extraInfo,
       'name': name,
     });
   }